+++ apps/web/src/features/auth/index.ts
@@ -1,5 +1,5 @@
-export { default as GoogleLoginButton } from './components/login-button/GoogleLoginButton';
-export { default as KakaoLoginButton } from './components/login-button/KakaoLoginButton';
-export { default as SocialLoginCallback } from './components/social-login-callback/SocialLoginCallback';
-export { default as TermsAgreementForm } from './components/terms-agreement/TermsAgreementForm';
-export { useTermsAgreement } from './apis/hooks/useTermsAgreement';
+export { useTermsAgreement } from '@/features/auth/apis/hooks/useTermsAgreement';
+export { default as GoogleLoginButton } from '@/features/auth/components/login-button/GoogleLoginButton';
+export { default as KakaoLoginButton } from '@/features/auth/components/login-button/KakaoLoginButton';
+export { default as SocialLoginCallback } from '@/features/auth/components/social-login-callback/SocialLoginCallback';
+export { default as TermsAgreementForm } from '@/features/auth/components/terms-agreement/TermsAgreementForm';
